Reorganize hero components, social links for mobile
