+++ package.json
@@ -8,6 +8,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@react-hookz/web": "^14.2.2",
     "next": "latest",
     "react": "17.0.2",
     "react-dom": "17.0.2"
